delete source check exist

+++ ajax/delete/source.php
@@ -5,13 +5,15 @@ $name = trim(strip_tags(filter_input(INPUT_POST, 'name', FILTER_DEFAULT)));
 $data['data'] = trim(strip_tags(filter_input(INPUT_POST, 'files', FILTER_DEFAULT)));
 $files = json_decode($data['data'], true);
 
-if($name && !empty($name)) {
+if ($name && !empty($name)) {
     foreach ($files as $i => $file) {
         if ($name === $file['name']) {
             unset($files[$i]);
             $data['data'] = empty($files) ? null : json_encode(array_values($files));
 
-            unlink(PATH_HOME . $file['url']);
+            if (file_exists(PATH_HOME . $file['url']))
+                unlink(PATH_HOME . $file['url']);
+
             break;
         }
     }
